Update README.md
When i needed to remove the daily limit by adding a google auth key, I had to decode the code and the google api to know that the key should be given in config, assigned to the param "key". It would have been helpful if it was mentioned in the doc.